Implement recusrive delimiting
Recursive delimting: instead of emitting all tokens as a linear stream,
they are emitted in the from of token trees. A token tree is a single
item in the token stream and can either be a recusive pair of () {} []
delimiters or a normal token.

This is adventageous, later when doing compile time metaprogramming as
for simpler meta programs the start and end of the invocation is already
prefdefined. Also for any delimiting mistakes, errors can alreday be
found during the lexing stage