[parser] Simplified parsing of `if`s (temporarily)
Ifs now always need to continue on with a block `{ ... }`,
in order to be valid. Simple if-expressions like `if cond code();`
are no longer valid as they are not uniquely parsable and deserable.

Maybe this issue will be fixed by creating an if-then expression for
those cases `if cond then code();`